pyproject.toml
[tool.poetry]
name = "yami-market-server"
version = "0.1.0"
description = "Yami Shopping Market"
authors = ["Zhen Jun Xu <me@arlenx.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
flask-cors = "^3.0.10"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.2"
flask-jwt-extended = "^4.4.4"
email-validator = "^1.3.0"
rich = "^12.6.0"
nanoid = "^2.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
isort = "^5.10.1"
mypy = "^0.991"
yapf = "^0.32.0"
toml = "^0.10.2"
pre-commit = "^2.20.0"
gunicorn = "^20.1.0"
types-psycopg2 = "^2.9.21.2"
types-flask-cors = "^3.0.10"
flake8-quotes = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 2
split_before_logical_operator = true

[tool.isort]
profile = "hug"

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
plugins = "pydantic.mypy"
